+++ service/controller/chart/resource/release/current_test.go
@@ -6,7 +6,7 @@ import (
 	"strconv"
 	"testing"
 
-	"github.com/giantswarm/apiextensions/v3/pkg/apis/application/v1alpha1"
+	"github.com/giantswarm/apiextensions-application/api/v1alpha1"
 	"github.com/giantswarm/helmclient/v4/pkg/helmclient"
 	"github.com/giantswarm/helmclient/v4/pkg/helmclienttest"
 	"github.com/giantswarm/micrologger/microloggertest"
@@ -141,7 +141,7 @@ func Test_CurrentState(t *testing.T) {
 
 			c := Config{
 				Fs:         afero.NewMemMapFs(),
-				CtrlClient: fake.NewFakeClient(),
+				CtrlClient: fake.NewFakeClient(), //nolint:staticcheck
 				HelmClient: helmClient,
 				K8sClient:  k8sfake.NewSimpleClientset(),
 				Logger:     microloggertest.New(),